% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_wrappers.R
\name{bash_col_sum}
\alias{bash_col_sum}
\title{A Function}
\usage{
bash_col_sum(f1, col)
}
\arguments{
\item{f1}{The name/path of the file.}

\item{col}{The name of the column to find the sum of.}
}
\description{
Finds the sum of a given column.
}
\examples{
bash_col_sum()
}
