% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_wrappers.R
\name{bash_head}
\alias{bash_head}
\title{Show Head Of File}
\usage{
bash_head(f1, nrows)
}
\arguments{
\item{f1}{The name/path of the file.}

\item{nrows}{The number of rows to show.}
}
\description{
Show the top n rows of a file.
}
