% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash-wrappers.R
\name{bash_join}
\alias{bash_join}
\title{Bash Join Function}
\usage{
bash_join(f1, col1, f2, col2, f_new)
}
\arguments{
\item{f1}{The name/path of the first file.}

\item{col1}{Column to use for joining from the first file.}

\item{f2}{The name/path of the second file.}

\item{col2}{Column to use for joining from the second file.}

\item{f_new}{Nme for new file.}
}
\description{
Join two files stored on disk.
}
