% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{causal_disc_to_daggity}
\alias{causal_disc_to_daggity}
\title{Identify Columns With A Percentage Missing}
\usage{
causal_disc_to_daggity(func, args)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Function to convert GSheets #NULL values and
empty strings to NAs.
}
