% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{change_range}
\alias{change_range}
\title{Change Range}
\usage{
change_range(x, new_min, new_max)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
