% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{check_key}
\alias{check_key}
\title{Recursive Function Builder}
\usage{
check_key(dict, key)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
