% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{cohens_h}
\alias{cohens_h}
\title{Init Project Function}
\usage{
cohens_h(p1, p2)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function sets up all I need to start a new project.
}
\examples{
mean_dp()
}
\keyword{cats}
