% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cols_missing_data}
\alias{cols_missing_data}
\title{Identify Columns With NAs}
\usage{
cols_missing_data(df, frac)
}
\arguments{
\item{df}{A dataframe}
}
\value{

}
\description{
Takes a dataframe and returns the columns which meet a threshold of NAs
}
