% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cor_interactive}
\alias{cor_interactive}
\title{Interactive Correlation Map}
\usage{
cor_interactive(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
An interactive correlation map of the numeric/integer variables
}
\description{
Produces an interactive correlation map of the numeric/integer variables
}
