% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{create_db_doc_template}
\alias{create_db_doc_template}
\title{create_db_doc_template}
\usage{
create_db_doc_template(con)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
