% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heat_map.R
\name{create_heat_map}
\alias{create_heat_map}
\title{Correlation Heatmap}
\usage{
create_heat_map(df, interactive = FALSE)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A correlation heatmap of the numeric columns of the dataframe.
}
\description{
Generate a statis or interactive heatmap.
}
