% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{create_iso8601_dates}
\alias{create_iso8601_dates}
\title{ISO-8601 Date Creation}
\usage{
create_iso8601_dates(vec)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
