% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requirements_script.r
\name{create_requirements_script}
\alias{create_requirements_script}
\title{Requirements Script}
\usage{
create_requirements_script()
}
\value{
Lists all loaded packages
}
\description{
hold
}
