% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{create_schema}
\alias{create_schema}
\title{Create Schema}
\usage{
create_schema(con, schema)
}
\arguments{
\item{con}{DB connection}

\item{schema}{Schema name}
}
\value{
The df variable without the anomalous columns
}
\description{
U
}
