% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{create_sql_insert_query}
\alias{create_sql_insert_query}
\title{CSV To SQL Insert Statement}
\usage{
create_sql_insert_query(df)
}
\arguments{
\item{con}{DB connection}

\item{schema}{Schema name}

\item{tbl_path}{The path to the table}
}
\value{
The df variable without the anomalous columns
}
\description{
g(iris)
}
