% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{create_sql_table_query}
\alias{create_sql_table_query}
\title{Get Create Table Query}
\usage{
create_sql_table_query(con, df, tbl_name)
}
\arguments{
\item{con}{DB connection}

\item{df}{schema Schema name}

\item{tbl_name}{The path to the table}
}
\value{
SQL query to create the dataframe as a table in a database
}
\description{
Get the SQL equivalent of creting a table from its tidyverse representation.
}
