% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{create_tbl}
\alias{create_tbl}
\title{TBL To DB}
\usage{
create_tbl(data, schema_name, tbl_name)
}
\arguments{
\item{con}{DB connection}

\item{schema}{Schema name}

\item{tbl_path}{The path to the table}
}
\value{
The df variable without the anomalous columns
}
\description{
U
}
