% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_train_test_split}
\alias{create_train_test_split}
\title{Combined Model Diagnostics}
\usage{
create_train_test_split(df, prob = 0.8)
}
\value{
Tidy model diagnostics with the confidence intervals of the estimates
}
\description{
Combines tidy model diagnostics with the confidence intervals of the estimates
}
