% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{date_decomposition}
\alias{date_decomposition}
\title{Date Decomposition}
\usage{
date_decomposition(df, col_name)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Given a date column this function separates the day, month, and year
into new columns
}
