% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{detect_outliers_mad}
\alias{detect_outliers_mad}
\title{detect_outliers_mad}
\usage{
detect_outliers_mad(group, interval = 2)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
