% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{dfs_equal}
\alias{dfs_equal}
\title{Dataframe Equality}
\usage{
dfs_equal(df1, df2)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
