% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{drop_anomalous_cols}
\alias{drop_anomalous_cols}
\title{Recursive Function Builder}
\usage{
drop_anomalous_cols(df)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
