% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{drop_na_col}
\alias{drop_na_col}
\title{Drop NA Columns}
\usage{
drop_na_col(df, method = "all", frac = 0.1)
}
\arguments{
\item{df}{A dataframe}

\item{method}{The drop criterion:
* all: Drop columns with all NAs
* frac: Drop columns with NAs above a threshold
* any: Drop any columns with at least one NA}

\item{frac}{The threshold (between 0 and 1 inclusive) to be used when method="frac".}
}
\value{
A dataframe with some columns dropped according to the supplied specification
}
\description{
Drops columns with NAs according to a specified criterion.
}
