% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{drop_na_col_some}
\alias{drop_na_col_some}
\title{Recursive Function Builder}
\usage{
drop_na_col_some(df, frac = 0.1, method = "all")
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
