% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{extract_json}
\alias{extract_json}
\title{Extract JSON}
\usage{
extract_json(df)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
