% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{freedman_diaconis_hist_bins}
\alias{freedman_diaconis_hist_bins}
\title{Freedman–Diaconis Histogram Binning}
\usage{
freedman_diaconis_hist_bins(vec)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\value{
The Freedman-Diaconis bins
}
\description{
Uses the Freedman–Diaconis rule for optimal histogram binning.
Also see: https://stats.stackexchange.com/questions/798/calculating-optimal-number-of-bins-in-a-histogram/862#862
}
