% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{get_adjustment_sets}
\alias{get_adjustment_sets}
\title{}
\usage{
get_adjustment_sets(dag, parent, child, effect = "direct")
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{

}
