% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{get_children}
\alias{get_children}
\title{Hold}
\usage{
get_children(dag, parent)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Hold
}
