% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_diagnostics}
\alias{get_model_diagnostics}
\title{Init Project Function}
\usage{
get_model_diagnostics(mdl)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function sets up all I need to start a new project.
}
\examples{
mean_dp()
}
\keyword{cats}
