% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{get_node}
\alias{get_node}
\title{Identify Columns With A Percentage Missing}
\usage{
get_node(formula_str, loc)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Function to convert GSheets #NULL values and
empty strings to NAs.
}
