% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{heat_map}
\alias{heat_map}
\title{Correlation Heatmap}
\usage{
heat_map(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A correlation heatmap of the numeric columns of the dataframe.
}
\description{
Correlation Heatmap
}
