% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{hist_fd_rule}
\alias{hist_fd_rule}
\title{Freedman–Diaconis Histogram Binning}
\usage{
hist_fd_rule(vec)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\value{
The Freedman-Diaconis bins
}
\description{
Uses the Freedman–Diaconis rule for optimal histogram binning
}
