% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{ipak}
\alias{ipak}
\title{Counts Rows Of On Disk File}
\usage{
ipak(pkg)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
A command line version of nrow:
https://gist.github.com/stevenworthington/3178163
ipak(c("ggplot2", "plyr", "reshape2"))
}
