% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{list_col_to_str}
\alias{list_col_to_str}
\title{Unpack List Col As String}
\usage{
list_col_to_str(df, col)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Function to convert GSheets #NULL values and
empty strings to NAs.
}
