% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lm_check}
\alias{lm_check}
\title{Linear Model Evaluation}
\usage{
lm_check(mdl)
}
\arguments{
\item{mdl}{A trained model}
}
\value{
A list of linear model evaluation metrics.
}
\description{
Calculates several metrics for a trained linear regression model
}
