% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ls_funcs}
\alias{ls_funcs}
\title{List Package Functions}
\usage{
ls_funcs(pkg)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Lists a package's functions as a dataframe.
}
