% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{make_formula}
\alias{make_formula}
\title{Rounded Mean Function}
\usage{
make_formula(target, additions, subtractions = NULL, powers = NULL)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function return the mean of vector to n decimal places while ignoring missing values.
}
\examples{
mean_dp()
}
