% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_tbl_to_df.r
\name{md_tbl_to_df}
\alias{md_tbl_to_df}
\title{Markdown Table to Dataframe}
\usage{
md_tbl_to_df(md_tbl)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Converts a markdown table to a dataframe.
}
