% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{min_max_scaler}
\alias{min_max_scaler}
\title{Min Max Scaler}
\usage{
min_max_scaler(vec, max_, min_)
}
\arguments{
\item{vec}{The vector to scale}

\item{max_}{The max of the scale}

\item{min_}{THe min of the scale}
}
\value{
The scaled vector
}
\description{
This function sets up all I need to start a new project.
}
