% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{mode_stats}
\alias{mode_stats}
\title{Mode}
\usage{
mode_stats(vec)
}
\arguments{
\item{vec}{A numeric/integer vector}
}
\value{
The mode of the submitted vector
}
\description{
Calculates the statistical mode
}
