% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{model_calibration}
\alias{model_calibration}
\title{Model Calibration}
\usage{
model_calibration(df, preds_col, binary_class_col)
}
\arguments{
\item{df}{A dataframe}

\item{preds_col}{A string with the name of the column containing the probabilistic
predictions}

\item{binary_class_col}{A string with the name of the column with the labels}
}
\description{
Model Calibration
}
