% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Combined Model Diagnostics}
\usage{
model_diagnostics(mdl)
}
\value{
Tidy model diagnostics with the confidence intervals of the estimates
}
\description{
Combines tidy model diagnostics with the confidence intervals of the estimates
}
