% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{move_col}
\alias{move_col}
\title{Move Column}
\usage{
move_col(df, col_name, pos)
}
\arguments{
\item{df}{A dataframe}

\item{col_name}{THe name of the column to move}

\item{pos}{The position to move the column to}
}
\description{
Move a dataframe's column to a given position.
}
