% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{na_conversion}
\alias{na_conversion}
\title{NULL to NA Conversion}
\usage{
na_conversion(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Function to convert GSheets #NULL values and
empty strings to NAs.
}
