% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{optimal_hist_bins}
\alias{optimal_hist_bins}
\title{Init Project Function}
\usage{
optimal_hist_bins(vec)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function sets up all I need to start a new project.
}
\examples{
mean_dp()
}
\keyword{cats}
