% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_iglewicz_hoaglin.r
\name{outliers_iglewicz_hoaglin}
\alias{outliers_iglewicz_hoaglin}
\title{Outliers In Small Samples}
\usage{
outliers_iglewicz_hoaglin(vec)
}
\arguments{
\item{vec}{A numeric/integer vector}
}
\value{
A boolean vector of which values are and are not outliers
}
\description{
A function to find outliers in small samples. Method seen here: https://bit.ly/2DFmsJr
}
