% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{partial_recursive}
\alias{partial_recursive}
\title{Recursive Function Builder}
\usage{
partial_recursive(func, args)
}
\arguments{
\item{func}{Name/path of the file}
}
\value{
A hierarchical partial function
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
