% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_csv_sample}
\alias{read_csv_sample}
\title{Read CSV Sample}
\usage{
read_csv_sample(fname, nrows, seed = 8, header = "-r")
}
\arguments{
\item{nrows}{The number of rows to sample}

\item{seed}{The seed to use for the random selection}

\item{header}{An indicator of whether the file has a header row. The default is yes ("=r"). Use "" for no.}

\item{fpath}{The pth and name of the file}
}
\value{
df A sample of the file specified
}
\description{
This function takes a path and reads all the csvs in it into a single dataframe.
}
