% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_csvs}
\alias{read_csvs}
\title{Read CSVs Function}
\usage{
read_csvs(path)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function takes a path and reads all the csvs in it into a single dataframe.
}
\examples{
mean_dp()
}
\keyword{cats}
