% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{read_db_data_thru_tunnel}
\alias{read_db_data_thru_tunnel}
\title{Read Data Thru Tunnel}
\usage{
read_db_data_thru_tunnel(qry)
}
\arguments{
\item{con}{DB connection}

\item{schema}{Schema name}

\item{tbl_path}{The path to the table}
}
\value{
The df variable without the anomalous columns
}
\description{
U
}
