% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mode.R
\name{read_mode}
\alias{read_mode}
\title{Mode}
\usage{
read_mode(vec)
}
\arguments{
\item{vec}{A numeric/integer vector}
}
\value{
The mode of the submitted vector
}
\description{
Calculates the statistical mode
}
