% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pkg_dependency_tree.r
\name{read_pkg_dependency_tree}
\alias{read_pkg_dependency_tree}
\title{Read Package Dependency Trees}
\usage{
read_pkg_dependency_tree(
  pack,
  dep_level = c("Depends", "Imports", "LinkingTo"),
  available_packages = available.packages()
)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
