% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{read_stratified_sampling_examples}
\alias{read_stratified_sampling_examples}
\title{Stratified Sampling}
\usage{
read_stratified_sampling_examples()
}
\value{
The df variable without the anomalous columns
}
\description{
U
}
