% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_wrappers.R
\name{remove_col_dups}
\alias{remove_col_dups}
\title{A Function}
\usage{
remove_col_dups(df)
}
\arguments{
\item{f1}{The name/path of the first file.}

\item{col1}{Column to use for joining from the first file.}

\item{f2}{The name/path of the second file.}

\item{col2}{Column to use for joining from the second file.}

\item{f_new}{Nme for new file.}
}
\description{
Join two files.
}
\examples{
bash_join()
}
