% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_cols.R
\name{remove_duplicate_cols}
\alias{remove_duplicate_cols}
\title{Remove Duplicated Columns}
\usage{
remove_duplicate_cols(df)
}
\arguments{
\item{df}{The dataframe to be modified.}
}
\value{
The data with duplicated columns removed
}
\description{
Generates unique columns for a dataframe.
}
