% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_py.R
\name{rmd_to_py}
\alias{rmd_to_py}
\title{Convert RMD To Python Script}
\usage{
rmd_to_py(input_file_path, output_file_path)
}
\value{
TRUE
}
\description{
Converts an Rmd with comments and Python chunks to a Python script
}
