% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{small_sample_outliers}
\alias{small_sample_outliers}
\title{Rounded Mean Function}
\usage{
small_sample_outliers(vec)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
https://stats.stackexchange.com/questions/78609/outlier-detection-in-very-small-sets/78617#78617
}
\examples{
mean_dp()
}
