% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{standard_scaler_df}
\alias{standard_scaler_df}
\title{Init Project Function}
\usage{
standard_scaler_df(x)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function sets up all I need to start a new project.
}
\examples{
mean_dp()
}
\keyword{cats}
