% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{time_series_clustering}
\alias{time_series_clustering}
\title{Time Series Clustering}
\usage{
time_series_clustering(mat, dist = "dtw", linkage = "average")
}
\arguments{
\item{mat}{A matrix}

\item{dist}{A distance measure. See the method arg in parallelDist::parDist for the supported parameters.}

\item{linkage}{The type of linkage to use. See the method arg in TSclust::hclust for the supported parameters.}
}
\value{
A hierarchical cluster object
}
\description{
Hierarchical clustering for time series with dynamic time warping.
}
