% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_test_split}
\alias{train_test_split}
\title{Train Test Split}
\usage{
train_test_split(df, frac)
}
\arguments{
\item{df}{A dataframe}

\item{frac}{The proportion of the data to be used for training}
}
\value{
A dataframe with an extra column called train where a 1 means the data point
is in the training set.
}
\description{
Splits a dataframe into training and testing sets
}
