% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ts_clustering}
\alias{ts_clustering}
\title{Time Series Clustering}
\usage{
ts_clustering(mat, dist = "dtw", linkage = "average")
}
\arguments{
\item{mat}{A matrix}

\item{dist}{A distance measure. See the method arg in parallelDist::parDist for the supported parameters.}

\item{linkage}{The type of linkage to use. See the method arg in TSclust::hclust for the supported parameters.}
}
\value{
The clustering results
}
\description{
This function sets up all I need to start a new project.
}
