% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{update_label_na}
\alias{update_label_na}
\title{Imputation Indicator Columns}
\usage{
update_label_na(df, col_with_na)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
