% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_tbl_to_db}
\alias{write_tbl_to_db}
\title{TBL To DB}
\usage{
write_tbl_to_db(con, schema, tbl_path)
}
\arguments{
\item{con}{DB connection}

\item{schema}{Schema name}

\item{tbl_path}{The path to the table}
}
\value{
The df variable without the anomalous columns
}
\description{
U
}
