% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{write_txt}
\alias{write_txt}
\title{Write Text File}
\usage{
write_txt(obj, fname)
}
\arguments{
\item{obj}{The object to write}

\item{fname}{The name of the file to write to}
}
\description{
Helper function to write an object to txt
}
