% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_wrappers.R
\name{bash_arrange}
\alias{bash_arrange}
\title{Column Sort}
\usage{
bash_arrange(f1, col_sort)
}
\arguments{
\item{f1}{The name/path of the file.}

\item{col_sort}{The name of the column to find the mean of.}
}
\description{
Sorting by a column
}
