% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash-wrappers.R
\name{bash_col_mean}
\alias{bash_col_mean}
\title{Mean}
\usage{
bash_col_mean(f1, col)
}
\arguments{
\item{f1}{The name/path of the file.}

\item{col}{The name of the column to find the mean of.}
}
\description{
Finds the mean of a given column.
}
