% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash-wrappers.R
\name{bash_ncols}
\alias{bash_ncols}
\title{Column Counter}
\usage{
bash_ncols(f1, sep = ",")
}
\arguments{
\item{f1}{Name/path of the file.}

\item{sep}{File separator. Defaults to csv.}
}
\value{
The number of columns
}
\description{
Returns the number of columns in a file
}
