% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_wrappers.R
\name{bash_select}
\alias{bash_select}
\title{A Function}
\usage{
bash_select(f1, cols_to_remove, f2 = NA)
}
\arguments{
\item{f1}{The name/path of the file.}

\item{cols_to_remove}{Names of the columns to removes.}

\item{f_new}{Optional name for new file if the results should be saved.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
bash_select()
}
