% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{causal_disc_to_daggity}
\alias{causal_disc_to_daggity}
\title{}
\usage{
causal_disc_to_daggity(func, args)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Causal Discovery To DAG: Recursively creates partial functions
to intake each formula except the last
}
