% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(dict, key)
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{The second dataframe}
}
\value{
TRUE if the dataframes are equal else FALSE
}
\description{
Checks if two dataframes are equal.
}
