% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_h.R
\name{cohens_h}
\alias{cohens_h}
\title{Cohen's H}
\usage{
cohens_h(p1, p2)
}
\arguments{
\item{p1}{The first probability or proportion}

\item{p2}{The second probability or proportion}
}
\description{
Calculates Cohen's H for two probabilities or proportions.
}
