% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{cols_missing_data}
\alias{cols_missing_data}
\title{Identify Columns With A Percentage Missing}
\usage{
cols_missing_data(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Function to convert GSheets #NULL values and
empty strings to NAs.
}
