% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cor_interactive}
\alias{cor_interactive}
\title{Init Project Function}
\usage{
cor_interactive(df)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function sets up all I need to start a new project.
}
\examples{
mean_dp()
}
\keyword{cats}
