% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{date_decomposition}
\alias{date_decomposition}
\title{Imputation Indicator Columns}
\usage{
date_decomposition(df)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
