% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{dfs_equal}
\alias{dfs_equal}
\title{Recursive Function Builder}
\usage{
dfs_equal(df1, df2)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
