% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{drop_anomalous_cols}
\alias{drop_anomalous_cols}
\title{Drop Anomalous Columns}
\usage{
drop_anomalous_cols(df)
}
\arguments{
\item{df}{A dataframe to be examined}
}
\value{
The df variable without the anomalous columns
}
\description{
Uses the xray library to identify anomalous columns and drops them.
}
