% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{get_df_adjustment_sets}
\alias{get_df_adjustment_sets}
\title{}
\usage{
get_df_adjustment_sets(dag)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
List all direct causal effects that are identifiable via backdoor adjustment
}
