% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_dummies}
\alias{get_dummies}
\title{Get Dummies}
\usage{
get_dummies(df, col)
}
\arguments{
\item{df}{A dataframe}

\item{col}{Column name to be dummied as string}
}
\value{
The dummified dataframe
}
\description{
Dummy a given categorical variable
}
