% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{impute_with_na}
\alias{impute_with_na}
\title{Replace With NA}
\usage{
impute_with_na(df, val = "")
}
\arguments{
\item{df}{A dataframe A numeric/integer vector}

\item{num_dp}{The value to replace by NA. The default is "".}
}
\value{
Data with NA in place of designated value
}
\description{
Replace a given value with NA
}
