% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{list_col_to_str}
\alias{list_col_to_str}
\title{Unpack List Column As String}
\usage{
list_col_to_str(df, col)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The name of the list column}
}
\value{
A dataframe with NAs instead of #NULLs and empty strings.
}
\description{
Converts a list column to a string column
}
