% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{make_formula}
\alias{make_formula}
\title{Dynamic Formula Construction}
\usage{
make_formula(target, additions, subtractions = NULL)
}
\arguments{
\item{target}{The LHS (left hand side) of the formula}

\item{additions}{The columns to add}

\item{substractions}{The columns to subtract}
}
\value{
A formula
}
\description{
A simple dynamic formula constructor
}
