% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_csv_sample}
\alias{read_csv_sample}
\title{Read CSVs Function}
\usage{
read_csv_sample(fpath, nrows, seed = 8, header = "-r")
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function takes a path and reads all the csvs in it into a single dataframe.
}
\examples{
mean_dp()
}
\keyword{cats}
