% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_csvs}
\alias{read_csvs}
\title{Read CSVs}
\usage{
read_csvs(path)
}
\arguments{
\item{path}{The path to the csvs to read in}
}
\description{
This function takes a path and reads all the csvs in it into a single dataframe.
}
