% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{remove_col_dups}
\alias{remove_col_dups}
\title{Remove Duplicated Columns}
\usage{
remove_col_dups(df)
}
\arguments{
\item{df}{The dataframe to be modified.}
}
\description{
Generates unique columns for a dataframe.
}
