% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ss_sampleseed}
\alias{ss_sampleseed}
\title{Seed Sampling}
\usage{
ss_sampleseed(df, frac, seed = 8)
}
\arguments{
\item{df}{A dataframe}

\item{frac}{The fraction}

\item{seed}{The seed for the sampling}
}
\description{
Produces a reproducible sample from a dataframe with
}
