% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_test_split}
\alias{train_test_split}
\title{Train Test Split}
\usage{
train_test_split(df, ratio)
}
\arguments{
\item{df}{The dataframe to splitr}

\item{ratio}{The fraction of data for training}
}
\value{
A list of the training and testing dataframes
}
\description{
This function sets up all I need to start a new project.
}
