% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{train_test_split}
\alias{train_test_split}
\title{Recursive Function Builder}
\usage{
train_test_split(df)
}
\arguments{
\item{f1}{Name/path of the file}
}
\description{
Recursively creates partial functions to intake each
formula except the last. For the last formula evaluation takes place
}
