% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{update_label_nas}
\alias{update_label_nas}
\title{Imputation Indicator Columns}
\usage{
update_label_nas(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with indicator columns for columns with at least one NA
}
\description{
For a given dataframe this functions adds a Boolean column for each column
with NAs that indicates the location of these NAs.
}
