% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{write_txt}
\alias{write_txt}
\title{Rounded Mean Function}
\usage{
write_txt(fobj, fname)
}
\arguments{
\item{col}{A numeric/integer vector}

\item{num_dp}{The numeric of decimal places to round the mean to.}
}
\description{
This function return the mean of vector to n decimal places while ignoring missing values.
}
\examples{
mean_dp()
}
